requirements.txt
pandas
request
BeautifulSoup